# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DDD asosida qurilgan e-commerce marketplace backend",
        "title": "Uzum Clone API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/dashboard/low-stock": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Eng kam zaxiraga ega (o'chirilmagan) top-5 mahsulot. Faqat admin uchun.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-dashboard"
                ],
                "summary": "Kam zaxirali mahsulotlar",
                "responses": {
                    "200": {
                        "description": "Kam zaxirali mahsulotlar",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dashboard.LowStockProduct"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/revenue-history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Daromadni davr (kun yoki oy) bo'yicha guruhlab qaytaradi, faqat to'langan va bekor qilinmagan buyurtmalar bo'yicha - frontend'dagi oddiy grafik uchun. Faqat admin uchun.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-dashboard"
                ],
                "summary": "Daromad tarixi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Guruhlash davri: day yoki month",
                        "name": "period",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daromad tarixi",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dashboard.RevenuePoint"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "period 'day' yoki 'month' bo'lishi kerak",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/admin/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Jami mahsulotlar soni (o'chirilganlardan tashqari), jami sotilgan birliklar va jami daromad (faqat to'langan va bekor qilinmagan buyurtmalar bo'yicha). Faqat admin uchun.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-dashboard"
                ],
                "summary": "Umumiy statistika",
                "responses": {
                    "200": {
                        "description": "Umumiy statistika",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dashboard.SummaryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya talab qilinadi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Faqat admin uchun",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Ichki server xatosi",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dashboard.LowStockProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name_uz": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "dashboard.RevenuePoint": {
            "type": "object",
            "properties": {
                "period": {
                    "type": "string"
                },
                "revenue": {
                    "type": "number"
                }
            }
        },
        "dashboard.SummaryResponse": {
            "type": "object",
            "properties": {
                "total_products": {
                    "type": "integer"
                },
                "total_revenue": {
                    "type": "number"
                },
                "total_units_sold": {
                    "type": "integer"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}